Add files via upload
This was a project was the winning entry for the iRobokids Corona safety, prevention and awareness coding competition held in April 2020. It was created by me when I was in grade 4, Billabong High International School, Malad. It is a two-player boss shooting game with bullets, enemies and loads of fun!