AJ cyber fiefdom
This is a retro-style board game recreated online where players move around using a dice on a board to complete tasks and gather or lose money. The one to get the maximum at the end wins the game! This game won the runner-up award in the design championship game development challenge 2021. I created this game with Jashith Narang. Have fun!